+++ lib/structure.rb
@@ -24,13 +24,18 @@ module Structure
     #   person.age  # => 30
     def new(&block)
       builder = Builder.new
-      builder.instance_eval(&block) if block
+      builder.instance_eval(&block) if block # steep:ignore
 
       # @type var klass: untyped
       klass = Data.define(*builder.attributes)
 
+      custom_methods_metadata = { instance: [], singleton: [] }.freeze # steep:ignore
+
       # Enable custom method definitions by evaluating block on the class
       if block
+        before_instance_methods = klass.instance_methods(false)
+        before_singleton_methods = klass.singleton_methods(false)
+
         # Provide temporary dummy DSL methods to prevent NoMethodError during class_eval
         klass.define_singleton_method(:attribute) { |*args, **kwargs, &blk| }
         klass.define_singleton_method(:attribute?) { |*args, **kwargs, &blk| }
@@ -43,6 +48,36 @@ module Structure
         klass.singleton_class.send(:remove_method, :attribute)
         klass.singleton_class.send(:remove_method, :attribute?)
         klass.singleton_class.send(:remove_method, :after_parse)
+
+        after_instance_methods = klass.instance_methods(false)
+        after_singleton_methods = klass.singleton_methods(false)
+
+        newly_defined_instance_methods = after_instance_methods - before_instance_methods
+        newly_defined_singleton_methods = after_singleton_methods - before_singleton_methods
+
+        instance_method_metadata = newly_defined_instance_methods.filter_map do |name|
+          next unless klass.public_method_defined?(name)
+
+          {
+            name: name,
+            parameters: klass.instance_method(name).parameters,
+          }.freeze
+        end
+
+        singleton_class = klass.singleton_class
+        singleton_method_metadata = newly_defined_singleton_methods.filter_map do |name|
+          next unless singleton_class.public_method_defined?(name)
+
+          {
+            name: name,
+            parameters: singleton_class.instance_method(name).parameters,
+          }.freeze
+        end
+
+        custom_methods_metadata = {
+          instance: instance_method_metadata.freeze,
+          singleton: singleton_method_metadata.freeze,
+        }.freeze
       end
 
       # Override initialize to make optional attributes truly optional
@@ -72,6 +107,7 @@ module Structure
         coercions: builder.coercions(klass),
         after_parse: builder.after_parse_callback,
         required: builder.required,
+        custom_methods: custom_methods_metadata,
       }.freeze
       klass.instance_variable_set(:@__structure_meta__, meta)
       klass.singleton_class.attr_reader(:__structure_meta__)