+++ test/fixtures/person.rb
@@ -2,7 +2,6 @@
 
 require "structure"
 
-# integer, boolean, simple predicate, optional attribute
 Person = Structure.new do
   attribute(:name, String)
   attribute?(:age, Integer)